code clean up and added a unit test

+++ src/main/java/com/vestas/libraryManagement/exception/BookNotFoundException.java
@@ -3,7 +3,7 @@ package com.vestas.libraryManagement.exception;
 import static java.lang.String.format;
 
 public class BookNotFoundException extends RuntimeException {
-    public BookNotFoundException(Long id) {
+    public BookNotFoundException(int id) {
         super(format("Book with id=%d not found", id));
     }
 }
